requirements.txt
-e git+https://github.com/NickNothom/pynmea2.git#egg=pynmea2
pyserial
pyinstaller
pywebview[qt,cef]